pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "douyin-xhs-mcp-server"
version = "1.3.0"
description = "MCP server for extracting content from Douyin and Xiaohongshu (videos, images, and text)"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "yzfly", email = "yz.liu.me@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "requests",
    "ffmpeg-python",
    "tqdm",
    "dashscope",
    "beautifulsoup4",
]

[project.urls]
Homepage = "https://github.com/yzfly/douyin-mcp-server"
Repository = "https://github.com/yzfly/douyin-mcp-server"
Issues = "https://github.com/yzfly/douyin-mcp-server/issues"

[project.scripts]
douyin-mcp-server = "douyin_mcp_server.server:main"
douyin-xhs-mcp-server = "douyin_mcp_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["douyin_mcp_server"]